+++ src/ska_low_mccs_spshw/station/tests/__init__.py
@@ -20,6 +20,7 @@ __all__ = [
     "TestRaw",
     "InitialiseStation",
     "BasicTangoTest",
+    "TestStationBeamDataRate",
 ]
 
 
@@ -30,6 +31,7 @@ from .test_channel_data import TestChannel
 from .test_integrated_beam import TestIntegratedBeam
 from .test_integrated_channel import TestIntegratedChannel
 from .test_raw_data import TestRaw
+from .test_station_beam_data_rate import TestStationBeamDataRate
 from .test_station_initialise import InitialiseStation
 from .test_tango import BasicTangoTest
 from .test_tile_beamformer import TestTileBeamformer