+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 * [SKB-702] SpsStation takes multiple attempts for all calibration coefficients to be applied
 * [THORN-5] Add Tile beamformer test to self-check.
 * [THORN-12] Add methods/attribute to measure data rate through DAQ network interface.
+* [THORN-13] Add station beam data rate test to self-check.
 
 ## 0.23.0